Test line

+++ backend/recipes/api_recipe.py
@@ -8,12 +8,14 @@ from starlette.background import BackgroundTask
 from starlette.requests import Request
 
 from db import database
-from recipes import schemas
 from recipes.models import (FavoriteCart, Ingredient, Recipe, Tag, cart,
                             favorites)
-from recipes.utils import (QueryParams, base64_image, create_cart_favorite,
-                           create_tag_ingredient, delete_cart_favorite)
-from services import query_list
+from recipes.schemas import (SFavorite, SIngredient, SIngredients, SlistRecipe,
+                             SLoadRecipe, SPatchRecipe, SRecipe, STag, STags)
+from recipes.utils import (QueryParams, base64_image, check_ingredient,
+                           check_tags, create_cart_favorite,
+                           delete_cart_favorite, utils_tag_ingredient)
+from services import image_delete, query_list
 from settings import FILES_ROOT, NOT_AUTHENTICATED, NOT_FOUND
 from users.models import User
 from users.utils import get_current_user
@@ -27,30 +29,42 @@ db_favorite_cart = FavoriteCart(database)
 
 
 @router.post("/tags/")
-async def create_tag(tag: schemas.Tag, user: User = PROTECTED):
-    return await create_tag_ingredient(db_tag.create_tag, tag, user)
+async def create_tag(tag: STag, user: User = PROTECTED):
+    """ Создавать теги может только администратор. """
+    return await utils_tag_ingredient(db_tag.create_tag, tag, user)
 
 
 @router.post("/ingredients/")
-async def create_ingredient(
-    ingredient: schemas.Ingredient, user: User = PROTECTED
-):
-    return await create_tag_ingredient(
+async def create_ingredient(ingredient: SIngredient, user: User = PROTECTED):
+    """ Создавать ингредиенты может только администратор. """
+    return await utils_tag_ingredient(
         db_ingredient.create_ingredient, ingredient, user)
 
 
-@router.get("/tags/", response_model=list[schemas.Tags])
-@router.get("/ingredients/", response_model=list[schemas.Ingredients])
-@router.get("/tags/{pk}/", response_model=schemas.Tags)
-@router.get("/ingredients/{pk}/", response_model=schemas.Ingredients)
+@router.post("/tag/{pk}/", response_model=STags)
+async def update_tag(tag: STag, pk: str, user: User = PROTECTED):
+    return await utils_tag_ingredient(db_tag.update_tag, tag, user, pk)
+
+
+@router.post("/ingredient/{pk}/", response_model=SIngredients)
+async def update_ingred(ingred: SIngredient, pk: int, user: User = PROTECTED):
+    return await utils_tag_ingredient(
+        db_ingredient.update_ingredient, ingred, user, pk)
+
+
+@router.get("/tags/", response_model=list[STags])
+@router.get("/ingredients/", response_model=list[SIngredients])
+@router.get("/tags/{pk}/", response_model=STags)
+@router.get("/ingredients/{pk}/", response_model=SIngredients)
 @router.delete("/tags/{pk}/")
 @router.delete("/ingredients/{pk}/")
-async def ingredients(
+async def tags_ingredients(
     request: Request,
     name: str = None,
     pk: int = None,
     user: User = PROTECTED
 ):
+    """ Удалять ингредиенты и теги может только администратор. """
     if "tags" in str(request.url.path):
         db_model_delete = db_tag.delete_tag
         db_model_get = db_tag.get_tags
@@ -59,31 +73,20 @@ async def ingredients(
         db_model_get = db_ingredient.get_ingredient
 
     if request.method == "DELETE":
-        if user.is_superuser or user.is_staff is True:
+        if user.is_staff or user.is_superuser:
             return await db_model_delete(pk)
         return Response(status_code=status.HTTP_403_FORBIDDEN)
     return await db_model_get(pk, name) or NOT_FOUND
 
 
-@router.post("/recipes/", response_model=schemas.Recipe)
+@router.post("/recipes/", response_model=SRecipe)
 async def create_recipe(
-    request: Request,
-    recipe: schemas.LoadRecipe,
-    user: User = PROTECTED
+    request: Request, recipe: SLoadRecipe, user: User = PROTECTED
 ):
     if not user:
         return NOT_AUTHENTICATED
-    for i in recipe.tags:
-        if not await db_tag.get_tags(i):
-            return JSONResponse(
-                {"detail": f"Not tag {i}"}, status.HTTP_404_NOT_FOUND
-            )
-    for i in recipe.ingredients:
-        i = i["id"]
-        if not await db_ingredient.get_ingredient(i):
-            return JSONResponse(
-                {"detail": f"Not ingredient {i}"}, status.HTTP_404_NOT_FOUND
-            )
+    await check_tags(recipe.tags)
+    await check_ingredient(recipe.ingredients)
     filename, image_path = await base64_image(recipe.image)
     try:
         recipe_item = {
@@ -96,14 +99,12 @@ async def create_recipe(
         pk = await db_recipe.create_recipe(
             recipe_item, recipe.ingredients, recipe.tags
         )
-        print("pk", pk)
         r = await db_recipe.get_recipe(pk, user_id=user.id, request=request)
         return r[0]
     except (Exception, UniqueViolationError):
-        if os.path.isfile(image_path):
-            os.remove(image_path)
+        await image_delete(image_path=image_path)
         return JSONResponse(
-            {"detail": "Error image"}, status.HTTP_400_BAD_REQUEST)
+            {"detail": "Error recipe"}, status.HTTP_400_BAD_REQUEST)
 
 
 @router.get("/recipes/download_shopping_cart/")
@@ -146,24 +147,31 @@ async def download_shopping_cart(user: User = PROTECTED):
     return FileResponse(file_path, background=BackgroundTask(cleanup))
 
 
-@router.get("/recipes/", response_model=schemas.listRecipe)
-@router.get("/recipes/{pk}/", response_model=schemas.Recipe)
+@router.get("/recipes/", response_model=SlistRecipe)
+@router.get("/recipes/{pk}/", response_model=SRecipe)
 @router.delete("/recipes/{pk}/", status_code=status.HTTP_204_NO_CONTENT)
 async def recipes(
     request: Request,
     pk: int = None,
     page: int = None,
     limit: int = None,
+    author: int = None,
     is_favorited: int = None,
     is_in_shopping_cart: int = None,
     user: User = PROTECTED,
     tags: QueryParams = Depends(QueryParams),
 ):
+    """
+    Получить все рецепты, один, удалить рецепт.
+    Администратор может удалять любые рецепты.
+    """
     user_id = user.id if user else None
     if request.method == "DELETE":
         if not user_id:
             return NOT_AUTHENTICATED
-        await db_recipe.delete_recipe(pk)
+        recipe = await db_recipe.check_recipe_by_id(pk)
+        if recipe.author_id == user.id or user.is_staff or user.is_superuser:
+            await db_recipe.delete_recipe(pk)
         return Response(status_code=status.HTTP_204_NO_CONTENT)
 
     if pk:
@@ -176,17 +184,56 @@ async def recipes(
 
     recipe_dict = await db_recipe.get_recipe(
         pk=pk, tags=tags, page=page, limit=limit, is_favorited=is_favorited,
-        user_id=user_id, is_in_cart=is_in_cart, request=request
+        user_id=user_id, author=author, is_in_cart=is_in_cart, request=request
     )
     if recipe_dict:
-        count = await db_recipe.count_recipe(tags, is_favorited, is_in_cart)
+        count = await db_recipe.count_recipe(
+            author, tags, is_favorited, is_in_cart
+        )
         return await query_list(recipe_dict, request, count, page, limit)
     return await query_list(recipe_dict, request, 0, page, limit)
 
 
-@router.post("/recipes/{pk}/favorite/", response_model=schemas.Favorite)
+@router.patch("/recipes/{pk}/", response_model=SRecipe)
+async def update_recipe(
+    request: Request, pk: int, recipe: SPatchRecipe, user: User = PROTECTED
+):
+    if not user:
+        return NOT_AUTHENTICATED
+    author = await db_recipe.check_recipe_image_by_id(pk)
+    if author.author_id == user.id or user.is_staff or user.is_superuser:
+        await check_tags(recipe.tags)
+        await check_ingredient(recipe.ingredients)
+        if recipe.image:
+            filename, image_path = await base64_image(recipe.image)
+            await image_delete(filename=author.image)
+        else:
+            filename, image_path = author.image, ""
+        try:
+            recipe_item = {
+                "author_id": author.author_id,
+                "name": recipe.name,
+                "text": recipe.text,
+                "image": filename
+            }
+            pk = await db_recipe.update_recipe(
+                pk, recipe_item, recipe.ingredients, recipe.tags
+            )
+            result = await db_recipe.get_recipe(
+                pk, user_id=user.id, request=request)
+            return result[0]
+
+        except UniqueViolationError as e:
+            await image_delete(filename, image_path)
+            return JSONResponse(
+                {"detail": e}, status.HTTP_400_BAD_REQUEST)
+
+    return Response(status_code=status.HTTP_403_FORBIDDEN)
+
+
+@router.post("/recipes/{pk}/favorite/", response_model=SFavorite)
 @router.delete("/recipes/{pk}/favorite/")
-@router.post("/recipes/{pk}/shopping_cart/", response_model=schemas.Favorite)
+@router.post("/recipes/{pk}/shopping_cart/", response_model=SFavorite)
 @router.delete("/recipes/{pk}/shopping_cart/")
 async def delete_cart(request: Request, pk: int, user: User = PROTECTED):
     if not user:
